NeuroDecodeR: A package for neural decoding analyses

The NeuroDecodeR makes it easy to do neural decoding analyses in R!

Details

The NeuroDecodeR (NDR) is built around five abstract object types that work together in a modular way to allow a range of neural decoding analyses. These five object types are:

Datasources (DS): Generate training and test splits of the data.

Feature preprocessors (FP): Learn parameters on the training set and apply transformations to the training and test sets.

Classifiers (CL): Learn the relationship between experimental conditions (i.e., "labels") and neural data on a training set, and then predict experimental conditions from neural data in a test set.

Result metrics (RM): Aggregate results across validation splits and over resampled runs and compute and plot final decoding accuracy metrics.

Cross-validators (CV): Take the DS, FP, CL and RM objects and run a cross-validation decoding procedure.

Data formats

Two data formats are used to do decoding analyses which are:

raster format contains high temporal precision data where neural activity from each site is stored in a separate file.

binned format contains data from multiple sites where the data is more coarsely binned across time.

A user of the NDR will typically store their data in raster format and then use the create_binned_data() to create a binned format data file that will be used in the decoding analysis.