A result metric (RM) method to aggregate results over cross-validation splits

This is a generic function that must be implemented by all RM objects. This function is called by the cross-validator results aggregated across all cross-validation splits. This method should not be called directly but instead is used internally by the cross-validator (CV) object.

aggregate_CV_split_results(rm_obj, prediction_results)

Arguments

rm_obj
The results metric object.

prediction_results
A data frame containing the prediction results to be aggregated over CV splits. The results in this data frame are the results returned by the CL's get_predictions() method, along with a column that specifies which cross-validation split the results come from. Thus the columns in the prediction_results data frame are: * CV: The cross-validation split number the results come from. * test_time: The time bin a test point comes from. * actual_labels: The actual labels for what happened on a trial. * predicted_labels: The predictions that classifier made. * decision_vals.___: A set of columns with the decision values for each class returned by the classifier.

Value

A result-metric object that contains the decoding results aggregated across cross-validation splits, and thus should take up less memory than the original prediction_results that was passed in to this method.

See also