Converts data from raster format to binned format

This function takes the name of a directory that contains files in raster format, and averages the data within a specified bin size at specifid sampling inverval increments to create data in binned format used for decoding.

create_binned_data(
  raster_dir_name,
  save_prefix_name,
  bin_width,
  sampling_interval,
  start_time = NULL,
  end_time = NULL,
  files_contain = ""
)

Arguments

raster_dir_name
A string that contains the path to a directory that has files in raster format. These files will be combined into binned format data.

save_prefix_name
A string with a prefix that will be used name of file that contains the saved binned format data.

bin_width
A number that has the bin width that data should be averaged over.

sampling_interval
A number that has the specifies the sampling interval between successive binned-data points.

start_time
A number that specifies at which time should the binning start. By default it starts at the first time in the raster data.

end_time
A number that specifies at which time should the binning end. By default is to end at the last time in the raster data.

files_contain
A string that specifies that only raster files that contain this string should be included in the binned format data.

Examples

# create binned data with 150 ms bin sizes sampled at 10 ms intervals
raster_dir_name <- "../data-raw/raster/Zhang_Desimone_7objects_raster_data_rda/"
# NOT RUN {
binned_file_name <- create_binned_data(raster_dir_name, "ZD", 150, 50)
# }