Saves the DECODING_RESULTS and logs the parameters used in the analysis

This function takes results returned by the cross-validator's run_decoding() method and uses the cross-validator's get_properties() method to save a log of the results that be used to reload the results.

log_save_results(
  DECODING_RESULTS,
  save_directory_name,
  result_name = "No result name set"
)

Arguments

DECODING_RESULTS

A list of results returned by the cross-validator's run_decoding method.

save_directory_name

A string specifying the directory name where the decoding results should be saved.

result_name

A string that gives a human readable name for the results that are to be saved. This name can be used to load the results later. The default value is "No result name set".