A plot function for the rm_main_results object

This function can create a line plot of the results or temporal cross-decoding results for the the zero-one loss, normalized rank and/or decision values after the decoding analysis has been run (and all results have been aggregated).

# S3 method for rm_main_results
plot(x, ..., results_to_show = "zero_one_loss", type = "TCD")

Arguments

x

A rm_main_result object that has aggregated runs from a decoding analysis, e.g., if DECODING_RESULTS are the out from the run_decoding(cv) then this argument should be DECODING_RESULTS$rm_main_results.

...

This is needed to conform to the plot generic interface.

results_to_show

A string specifying the types of results to plot. Options are: 'zero_one_loss', 'normalized_rank', 'decision_values', or 'all'.

type

A string specifying the type of results to plot. Options are 'TCD' to plot a temporal cross decoding matrix or 'line' to create a line plot of the decoding results as a function of time.

Value

Returns a ggplot object that plots the main results.

See also