A result metric (RM) that calculates main decoding accuracy measures

This result metric calculate the zero-one loss, the normalized rank, and the mean of the decision values. This is also an S3 object which has an associated plot function to display the results.

rm_main_results(
  ndr_container_or_object = NULL,
  aggregate_decision_values = TRUE,
  aggregate_normalized_rank = TRUE
)

Arguments

ndr_container_or_object
The purpose of this argument is to make the constructor of the rm_main_results feature preprocessor work with the magrittr pipe (%>%) operator. This argument should almost never be directly set by the user to anything other than NULL. If this is set to the default value of NULL, then the constructor will return a rm_main_results object. If this is set to an ndr container, then a rm_main_results object will be added to the container and the container will be returned. If this argument is set to another ndr object, then both that ndr object as well as a new rm_main_results object will be added to a new container and the container will be returned.

aggregate_decision_values
A string or boolean specifying how the decision values should be aggregated. If this is a boolean set to TRUE or to the string "full", then the decision values for the correct category will be calculated. If this is a boolean set to FALSE or to the string "none", then the decision values will not be calculated. If this is a string set to either "diag" or "only same train test time" then the decision values will only be calculated when for results when training and testing at the same time. Not returning the full results can speed up the runtime of the code and will use less memory so this can be useful for large data sets.

aggregate_normalized_rank
A string or boolean specifying how the normalized rank results should be aggregated. If this is a boolean set to TRUE or to the string "full", then the decision values for the correct category will be calculated. If this is a boolean set to FALSE or to the string "none", then the decision values will not be calculated. If this is a string set to either "diag" or "only same train test time" then the decision values will only be calculated when for results when training and testing at the same time. Not returning the full results can greatly speed up the runtime of the code and will use less memory so this can be useful for large data sets.

Details

Like all result metrics, this result metric has functions to aggregregate results after completing each set of cross-validation classifications, and also after completing all the resample runs. The results should then be available in the DECODING_RESULTS object returned by the cross-validator.

See also

Examples

# If you only want to use the rm_main_results(), then you can put it in a
# list by itself and pass it to the cross-validator.
the_rms <- list(rm_main_results())