NDTr: A package for neural decoding analyses

The NDTr is a package that makes it easy to do neural decoding analyses.

Details

The NDTr is built around 5 abstract object types that work together in a modular way to allow a range of neural decoding analyses. These five object types are:

Datasources (DS): Generate training and test splits of the data.

Feature preprofessors (FP): Learn parameters on the training set and apply transformations to the training and test sets.

Classifiers (CL): Learn the relationship between experimental conditions (i.e., "labels") and neural data on a training set, and then predict experimental conditions neural data in a test set.

Result metrics (RM): Aggregate results across validation splits and over resampled runs and compute and plot final decoding accuracy metrics.

Cross-validators (CV): Take the DS, FP, CL and RM objects and run a cross-validation decoding procedure.

Data formats

Two data formats are used to do decoding analyses which are:

raster format contains high temporal precision data where neural activity from each site is stored in a spearate file.

binned format contains data from multiple sites and the data is more coarsely binned across time.

A user of the NDTr will typically store their data in raster format and then use the create_binned_data() to create a binned format data file that will be used in the decoding analysis.