A function that converts data from raster format to binned format

This function takes the name of a directory that contains files in raster format, and averages the data within a specified bin size at specifid sampling inverval increments to create data in binned format used for decoding.

create_binned_data(raster_dir_name, save_prefix_name, bin_width,
  sampling_interval, start_time = NULL, end_time = NULL,
  files_contain = "")

Arguments

raster_dir_name
A string that contains the path to a directory that has files in raster format. These files will be combined into binned format data.

save_prefix_name
A string with a prefix that will be used name of file that contains the saved binned format data.

bin_width
A number that has the bin width that data should be averaged over.

sampling_interval
A number that has the specifies the sampling interval between successive binned-data points.

start_time
A number that specifies at which time should the binning start. By default it starts at the first time in the raster data.

end_time
A number that specifies at which time should the binning end. By default is to end at the last time in the raster data.

files_contain
A string that specifies that only raster files that contain this string should be included in the binned format data.

Examples

# create binned data with 150 ms bin sizes sampled at 10 ms intervals
raster_dir_name <- file.path('..', '..', 'data', 'raster',
                             'Zhang_Desimone_7objects_raster_data_rda')

binned_file_name <- create_binned_data(raster_dir_name, "ZD", 150, 50)

#>  binning site 1     binning site 2     binning site 3     binning site 4     binning site 5     binning site 6     binning site 7     binning site 8     binning site 9     binning site 10    binning site 11    binning site 12    binning site 13    binning site 14    binning site 15    binning site 16    binning site 17    binning site 18    binning site 19    binning site 20    binning site 21    binning site 22    binning site 23    binning site 24    binning site 25    binning site 26    binning site 27    binning site 28    binning site 29    binning site 30    binning site 31    binning site 32    binning site 33    binning site 34    binning site 35    binning site 36    binning site 37    binning site 38    binning site 39    binning site 40    binning site 41    binning site 42    binning site 43    binning site 44    binning site 45    binning site 46    binning site 47    binning site 48    binning site 49    binning site 50    binning site 51    binning site 52    binning site 53    binning site 54    binning site 55    binning site 56    binning site 57    binning site 58    binning site 59    binning site 60    binning site 61    binning site 62    binning site 63    binning site 64    binning site 65    binning site 66    binning site 67    binning site 68    binning site 69    binning site 70    binning site 71    binning site 72    binning site 73    binning site 74    binning site 75    binning site 76    binning site 77    binning site 78    binning site 79    binning site 80    binning site 81    binning site 82    binning site 83    binning site 84    binning site 85    binning site 86    binning site 87    binning site 88    binning site 89    binning site 90    binning site 91    binning site 92    binning site 93    binning site 94    binning site 95    binning site 96    binning site 97    binning site 98    binning site 99    binning site 100   binning site 101   binning site 102   binning site 103   binning site 104   binning site 105   binning site 106   binning site 107   binning site 108   binning site 109   binning site 110   binning site 111   binning site 112   binning site 113   binning site 114   binning site 115   binning site 116   binning site 117   binning site 118   binning site 119   binning site 120   binning site 121   binning site 122   binning site 123   binning site 124   binning site 125   binning site 126   binning site 127   binning site 128   binning site 129   binning site 130   binning site 131   binning site 132