A function that converts data from raster format to binned format

This function takes takes a directory containing data in raster format and converts it into a file that contains 'binned' spike count rate data that is sampled at a particular frequency.

create_binned_data(raster_directory_name, save_prefix_name, bin_width, sampling_interval, start_ind, end_ind, files_contain)

creates a binned data file.

create_binned_data(raster_directory_name, save_prefix_name, bin_width,
  sampling_interval, start_ind = NULL, end_ind = NULL, files_contain = "")

Arguments

raster_directory_name
character. This argument is the name of a directory that contains files that are in raster format.

save_prefix_name
integer. This argument contains a prefix (e.g., directory name and additional characters) for how the resulting binned file should be saved. The string _binned_data_Xms_bins_Yms_sampled is appended on to the saved Rda file name, where X is the bin_width and Y is the sampling interval.

bin_width
integer. The bin width over which spike count rates should be computed.

sampling_interval
integer. The frequency with which the data should be sampled.

start_ind
integer. An optional argument that specifies the time index where the the first bin should start.

end_ind
logical. An optional argument that specifies the time index where the the binning process should end.

files_contain
character. Only raster data files that contain the file_contains string will be included in the binned data.

Examples

Bin the data using 150 ms bins sample at 50 ms intervals
Assumes that raster files are in the directory '../data/Zhang_Desimone_7objects_R_raster_data/' 
and saves the output file with the prefix ZD

#> Error: <text>:1:5: unexpected symbol
#> 1: Bin the
#>         ^
# NOT RUN {
create_binned_data('../data/Zhang_Desimone_7objects_R_raster_data/', '../data/ZD', 150, 10)

# all options
create_binned_data(raster_directory_name, save_prefix_name, bin_width, sampling_interval, start_ind, end_ind, files_contain)

# }