A plot function for data in raster format

This function will plot data that is in raster format. If the data is a spike train consisting of only 0's and 1's then it will create a plot of spikes as black tick marks on a white background. If the raster data contains continuous data, then the plot will be color coded.

# S3 method for raster_data
plot(x, ..., facet_label = NULL)

Arguments

x
Either data that is in raster format, or a string containing the name of a file that has data in raster format.

...
This is needed to conform to the plot generic interface.

facet_label
If this is set to a string that is the name of one of the labels in the raster data, then the raster plots will be faceted by this label.